Iceberg REST Service: Throw 400 if namespace is invalid (#17)
Fixes #10 to have the right behavior of a 400 when failing.

Currently, an IllegalArgumentException for invalid namespace will be
thrown which will end up surfacing as a 500 response due to
https://github.com/unitycatalog/unitycatalog/blob/main/server/src/main/java/io/unitycatalog/server/exception/GlobalExceptionHandler.java#L26
treating this as an arbitrary runtime exception. Instead a Base
exception with an invalid argument error code should be thrown (looking
at other parts of the code, this seems to be the practice).