Introduce AssertJ for testing
This introduces the AssertJ library for a more fluent way of testing and asserting things.
This library is widely used and improves reading and testing code (it's also being used in the Iceberg codebase).

I took the libery to convert the `IcebergRestCatalogTest` to AssertJ assertions to show its fluent nature.