EXTERNAL table creation for existing directory (#72)
**PR Checklist**

- [x] A description of the changes is added to the description of this
PR.
- [x] If there is a related issue, make sure it is linked to this PR.
- [ ] If you've fixed a bug or added code that should be tested, add
tests!
- [ ] If you've added or modified a feature, documentation in `docs` is
updated

**Description of changes**

This is related to the preprocessing step of EXTERNAL table creation
using CLI for local paths. Currently the CLI was checking if the
directory does not exist, before proceeding to create a delta table log
at the location. This was the only case for which EXTERNAL table
creation succeeded.

Ideally, new EXTERNAL table creation should succeed even for the case
when there's a delta log already present at the location (although we
should check if the schema confirms). If the directory does not exist or
there's no delta log at the location, it should create one. The current
PR implements the same.

The PR takes care of the
[issue](https://github.com/unitycatalog/unitycatalog/issues/62)

---------

Co-authored-by: Ramesh Chandra <ramesh.chandra@gmail.com>