implement proper error responses for Iceberg REST endpoints (#85)
**PR Checklist**

- [x] A description of the changes is added to the description of this
PR.
- [x] If there is a related issue, make sure it is linked to this PR.
- [x] If you've fixed a bug or added code that should be tested, add
tests!
- [ ] If you've added or modified a feature, documentation in `docs` is
updated

**Description of changes**

Added a specific Iceberg REST handler for the Iceberg REST
service/endpoints. This produces JSON error responses in line with
Iceberg REST spec.

Also stubbed out an endpoint for loadView because the Iceberg REST
client tries to load a table with given name and if table is not found,
it will try to load a view with given name (so, for now, we're just
returning a 404 for that endpoint).

**Related Issue**

Addresses https://github.com/unitycatalog/unitycatalog/issues/46

**Testing**

Added Iceberg REST error parsing to current tests.