add support for tables located in S3 in Iceberg REST catalog (#123)
**PR Checklist**

- [x] A description of the changes is added to the description of this
PR.
- [x] If there is a related issue, make sure it is linked to this PR.
- [x] If you've fixed a bug or added code that should be tested, add
tests!
- [ ] If you've added or modified a feature, documentation in `docs` is
updated

**Description of changes**

This PR introduces the ability to load Iceberg tables that are located
in S3. (Previously only local path was supported). This adds new
dependencies on iceberg-aws module as well as aws sdk v2. This also
introduces new testing frameworks - mockito and S3Mock in order to
better test and simulate reading an iceberg metadata out of S3 with
S3FileIO (as well as starting to use JUnit5 instead of JUnit4).

**Related Issues**

This PR covers the basic and initial S3 support addressing issue
https://github.com/unitycatalog/unitycatalog/issues/105