docs: add Delta Lake page (#246)
**Description of changes**

This PR adds a page explaining Delta Lake architecture and usage. Will
close #195

TBD: where to place this page in the docs. Options:

a. where it is now, maybe with a different title ("How Delta tables
work")
b. under `tables/...` although this is not so much about usage and more
about architecture
c. under a new heading `architecture/...`

open to hearing from folks on this!