Run Unity Catalog in a Docker container (#116)
# Description of changes

For this PR I took an alternative approach to #18 and #22 and created a
Dockerfile with build and start scripts that require minimal
intervention and interaction with the codebase.

From the codebase the only change is the .gitignore where I added the
.DS_Store which can be helpful in the future for contributors using Mac
OS,

PR #22 is a great start but maybe oversimplified.

PR #18 Has good thought put into it but I wanted to stay close to the
recommended way of running Unity Catalog as outlined in the project's
README. I tried not to fiddle directly with the jars and use the
provided `/bin/start-uc-server` to run the catalog. With this approach
the Dockerfile remains focused on building the environment and any
changes to how the environment should run can be made in the future
inside the `start-uc-server` script rather than the Dockerfile.

# Rationale of the PR

This pull request introduces a way to run Unity Catalog using Docker
containers. It provides a Dockerfile that builds the necessary
environment and separate bash scripts for building and starting the
catalog. This simplifies the process for users by requiring minimal
interaction with the codebase itself. The included README provides
detailed instructions on how to use these scripts to build and run the
Unity Catalog container.

> [!NOTE]
> The `README.md` contains two API calls that create an external and an
managed table.
> These APIs are not working yet because they are not supported by the
catalogue yet.

Signed-off-by: Jean Boutros <jean.boutros@gmail.com>

---------

Signed-off-by: Jean Boutros <jean.boutros@gmail.com>
Co-authored-by: Fokko Driesprong <fokko@apache.org>
Co-authored-by: Denny Lee <denny.g.lee@gmail.com>
Signed-off-by: Kevin Wang <kevin@eventualcomputing.com>