adjust --help messages in uc cli (#327)
**PR Checklist**

- [x] A description of the changes is added to the description of this
PR.
- [x] If there is a related issue, make sure it is linked to this PR.
- [x] If you've fixed a bug or added code that should be tested, add
tests!
- [x] If you've added or modified a feature, documentation in `docs` is
updated

**Description of changes**

Hi,

This is my first PR, so I would like to say hello and highlight that
this PR probably needs some attention. :)

Changes:

1. Added a new entry, `CliParams`, to distinguish between
`schema_full_name` for schema (which consists of two levels:
`catalog.schema`) and 'full_name" for other objects
(table/function/schema with a three-level name like
`catalog.schema.object`).
2. `printSubCommandHelp` now prints common parameters, including
`--output`, `--server`, and `--token`.
3. Minor fixes to messages used in `--help`.

The code in this PR changes `full_name` to `schema_full_name`, which may
be a breaking change. We need to consider whether this is the best
option or if we should stick to `full_name` in all cases and find an
alternative solution.

Issue: [253](https://github.com/unitycatalog/unitycatalog/issues/253)