Create GitHub workflow to assign issues to users on comment 'take' (#188)
**PR Checklist**

- [x] A description of the changes is added to the description of this
PR.
- [x] If there is a related issue, make sure it is linked to this PR
(close #121).
- [ ] If you've fixed a bug or added code that should be tested, add
tests!
- [x] If you've added or modified a feature, documentation in `docs` is
updated

**Description of changes**
This PR is related to #121

* Add auto-assign functionality for issues based on comments
- **CONTRIBUTING.md**: Add a section about auto-assigning issues based
on comments. Mention that users can comment "take" on an issue to be
auto-assigned the issue.
- **.github/workflows/auto_assign_issues.yml**: Create a new workflow
file named `auto_assign_issues.yml`. Add a trigger for issue comments.
Check if the comment is "take". Assign the issue to the user who
commented "take". Check if the user is already assigned to the issue and
log a message if the issue cannot be assigned.

**Further comments**
* I have already tested this issue. [Here's the link to
it.](https://github.com/nijanthanvijayakumar/unitycatalog/issues/2)
* This workflow is inspired by
https://github.com/MrPowers/quinn/blob/main/.github/workflows/assign-on-comment.yml
and the same has already been mentioned in the workflow file.
* I have updated the CONTRIBUTING.md (docs) file as well to reflect this
functionality is available to the users.

@MrPowers and @filipovskid - requesting a review.

---------

Signed-off-by: Tathagata Das <tathagata.das1565@gmail.com>
Co-authored-by: Tathagata Das <tathagata.das1565@gmail.com>
Signed-off-by: Kevin Wang <kevin@eventualcomputing.com>