Simple Token based access control and Oauth2 authentication (#277)
**PR Checklist**

- [x] A description of the changes is added to the description of this
PR.
- [x] If there is a related issue, make sure it is linked to this PR.
- [ ] If you've fixed a bug or added code that should be tested, add
tests!
- [ ] If you've added or modified a feature, documentation in `docs` is
updated

**Description of changes**

This PR adds (optional) authentication and access control to the UC
server. This is the first in a series to support Issue
https://github.com/unitycatalog/unitycatalog/issues/229

**Access control**
Access control is implemented using JWT Bearer tokens that are generated
and verified using internal signing keys. With authorization enabled,
clients accessing the UC must include an `Authorization` header with a
`Bearer` token that is generated through a token exchange with the UC.
When using the `uc` command, for example, an access token can be
provided with the command line option `--auth_token`. (See the
*Authentication* section for information on how to obtain access tokens)
```
bin/uc catalog list --auth_token <token>
```

The signing keys for creating access tokens is randomly generated on
startup and stored in `etc/conf/` if they do not exist. These files are
- `private-key.der` - an RSA512, 2048-bit private key, used for token
signing and validation
- `public-key.der` - a matching RSA512 public key, used for token
signing and validation
- `key-id.txt` - a random 256-bit id, used for JWKS verification
- `certs.json` - a representation of all the previous files for JWKS
verification

The server also generates an initial access token which is also stored
in `etc/conf/`.
- `token.txt` - the initial JWT access token

Enabling access control is done in the `server.properties` file. The
property
- `server.authorization` - controls whether access control is required
or not (values are `disable` or `enable`)

**Authentication**
Authentication is implemented via OAuth2 token exchange. This allows a
client to submit an identity token, from an identity provider like
Google or Okta, and exchange it for a server access token.

A new endpoint for this functionality
- `/api/2.1/unity-catalog/auth/tokens` - an OAuth2 token exchange
endpoint

This endpoint accepts an identity token, along with other OAuth2
parameters, validates the token using JWKS and OIDC discovery, and if
validation is successful, generates and returns a new JWT access-token
(generated using the previously described signing parameters)

_Token exchange with `uc` command_
The `uc` command has been enhanced with an `auth` `login` sub-command,
in which you provide an identity token with the `--identity_token`
parameter. As an example, you can obtain a Google Cloud identity token
with
```
gcloud auth print-identity-token
```
And then perform the token exchange with
```
bin/uc auth login --identity_token <token>
```

_Identity provider login with `uc` command_
The `uc` command can also do a full OAuth2 Authorization Flow with
identity providers like Google, Okta and Keycloak. In order to do that,
there are additional parameters to set in the `server.properties` file
- `server.authorization-url` - The authorization URL for the identity
provider
- `server.token-url` - The token URL for the identity provider
- `server.client-id` - The identity provider account Issuer client id
- `server.client-secret` - The identity provider account issuer client
secret
- `server.redirect-port` - Optional port number to use for the redirect
phase (Okta needs this)

With these properties configured, use the same command
```
bin/uc auth login
```
*without* the `--identity_token` parameter to initiate and Authorization
Flow and token exchange with Google.

**Compute engine authentication**

_Duckdb_

The only change you need to make is to supply a valid access-token when
configurating the secret:
```
CREATE SECRET (
      TYPE UC,
      TOKEN '<access-token>',
      ENDPOINT 'http://127.0.0.1:8080',
      AWS_REGION 'us-east-2'
 );
```

_Trino_

With Trino, you just need to supply an access-token when configuring the
catalog `catalog/iceberg.properties`:
```
connector.name=iceberg
iceberg.catalog.type=rest
iceberg.rest-catalog.uri=http://127.0.0.1:8080/api/2.1/unity-catalog/iceberg
iceberg.rest-catalog.security=OAUTH2
iceberg.rest-catalog.oauth2.token=<access-token>
```