[DOCKER] Use assembly jars for UC images (#367)
As reported in https://github.com/unitycatalog/unitycatalog/issues/339,
Docker containers fail to start after a successful docker image build.

This is due to
[this](https://github.com/unitycatalog/unitycatalog/commit/119f4707d033add34c0894916f544785e110fe94)
and
[this](https://github.com/unitycatalog/unitycatalog/commit/009ac77e7cd7926acb061ed1b8b5c3e5177c1bf7)
commits that introduced assembly jars with incomplete metadata
(`mainClass`).

This PR is supposed to fix
https://github.com/unitycatalog/unitycatalog/issues/339 with some extra
(opinionated) changes like:

1. UC CLI container accepts regular CLI options
2. No more `network` and `volume` (which may or may not be useful but
turned out a nightmare to make right to achieve the above)

There are some FIXMEs left to work on in followup PRs 😉

Signed-off-by: Kevin Wang <kevin@eventualcomputing.com>