[README] Add increasing JVM heap size in commands for running tests (#386)
**PR Checklist**

- [x] A description of the changes is added to the description of this
PR.
- [x] If there is a related issue, make sure it is linked to this PR.
- [ ] If you've fixed a bug or added code that should be tested, add
tests!
- [ ] If you've added or modified a feature, documentation in `docs` is
updated

**Description of changes**

Heap size when running tests was increased in the CI workflow in PR #274
Without this increase, the example commands for running tests given in
the README fail with `java.lang.OutOfMemoryError: Java heap space`.

This PR adds the same heap size increase to the commands for running
tests in the README.

Fixes #385

Signed-off-by: Kevin Wang <kevin@eventualcomputing.com>