Add a user-database and constrain access to known users. (#387)
**PR Checklist**

- [X] A description of the changes is added to the description of this
PR.
- [x] If there is a related issue, make sure it is linked to this PR.
- [x] If you've fixed a bug or added code that should be tested, add
tests!
- [ ] If you've added or modified a feature, documentation in `docs` is
updated

**Description of changes**

This enhancement adds a user database to the platform, a control API for
user management, CLI commands to manage users, and when authentication
is enabled, will constrain access to only the users in the user
database.

Related Issue https://github.com/unitycatalog/unitycatalog/issues/229

The user record is simple for this first implementation
- id - unique identifier for the user
- name - The user's full name
- email - The user's primary email address
- picture_url - A link to an avatar for the user
- external_id - An identifier for external identity providers to track
the user
- created_at - the time the record was created
- updated_at - the last time the record was updated

There is the beginning of a Control API which is meant to house
operations that are outside of the core of the Unity Catalog
specification.

The Control API initially provides endpoints to do user management. The
endpoints are SCIM 2.0 compatible, as such they should allow integration
with third party identity providers for syncing users between systems.

The CLI enhancements add commands to add, update and remove users, see
command line help for details, but as an example, to add a user
```
uc user create --name "Iggy Pop" --email iggy@thestooges.com
```

Finally, when the recently added OAuth authorization functionality is
enabled, users must exist in the user database in order to gain access
to the server operations.

---------

Signed-off-by: Mocker <mock@fakebelieve.org>
Co-authored-by: Xiang Xu <xiang.xu@databricks.com>
Co-authored-by: Ramesh Chandra <ramesh.chandra@gmail.com>
Signed-off-by: Kevin Wang <kevin@eventualcomputing.com>