Fix the `/self` endpoint. (#428)
Add some checking in currentUser() for the token before accessing it.

**PR Checklist**

- [x] A description of the changes is added to the description of this
PR.
- [x] If there is a related issue, make sure it is linked to this PR.
- [x] If you've fixed a bug or added code that should be tested, add
tests!
- [x] If you've added or modified a feature, documentation in `docs` is
updated

**Description of changes**

The `/self` endpoint relies on the acess-token JWT being in the request
context. In order to do that, it needs to be intercepted by the
`AuthDecorator` first, so adding that to the request pipeline, excluding
`auth/tokens` which should be open.

Also made some changes to `/self` to make sure it found the token before
trying to access it.

And the token exchange verifies the incoming token subject is in the
user database.

Issue https://github.com/unitycatalog/unitycatalog/issues/429

Signed-off-by: Kevin Wang <kevin@eventualcomputing.com>