[DOCKER] UC server container displays welcome message (#392)
This PR changes how the UC server docker image is built (uses the faster
`sbt server/package`) and copies the jars required to start it (to
`jars` directory).

After
https://github.com/unitycatalog/unitycatalog/commit/4b7aaa319bfad76dde4d30c9fdd5128895bf98d1,
I thought I'd work on changes that would make the behaviour of the UC
server when started consistent regardless whether it's in or outside
Docker container (namely the welcome message at the start and no
warnings).

A Dockerized UC runs in the foreground and containers are removed once
stopped.

Signed-off-by: Kevin Wang <kevin@eventualcomputing.com>