Add python sdk (#589)
**PR Checklist**

- [x] A description of the changes is added to the description of this
PR.
- [x] If there is a related issue, make sure it is linked to this PR.
- [x] If you've fixed a bug or added code that should be tested, add
tests!
- [x] If you've added or modified a feature, documentation in `docs` is
updated

**Description of changes**

This PR adds a Python client SDK generated from the OpenAPI spec.

Unlike
[unitycatalog/unitycatalog-python](https://github.com/unitycatalog/unitycatalog-python),
it does not have a vendor dependency with Stainless but instead uses the
[openapi-generator](https://github.com/OpenAPITools/openapi-generator/)
library. As a result, the library interface is different.

This PR adds a step to `build.sbt`, which runs when `build/sbt generate`
is called and generates a Python client at `target/clients/python`.

Follow-up from #309, resolves #286

---------

Signed-off-by: Kevin Wang <kevin@eventualcomputing.com>
Signed-off-by: abhinavmittal554 <abhinavmittal554@gmail.com>