Add usage warnings for poorly constructed functions (#690)
**PR Checklist**

- [X] A description of the changes is added to the description of this
PR.
- [ ] If there is a related issue, make sure it is linked to this PR.
- [ ] If you've fixed a bug or added code that should be tested, add
tests!
- [ ] If you've added or modified a feature, documentation in `docs` is
updated

**Description of changes**

Introduce a warning for externally created UC functions that will be
used as a GenAI tool to inform users that a lack of a function
description or parameter descriptions severely impacts the viability of
using UC functions as GenAI tools if these are not defined.
There are no updates to be made for functions created with the
unitycatalog-ai `create_python_function` as this is already not
permitted within that API.

---------

Signed-off-by: Ben Wilson <benjamin.wilson@databricks.com>
Signed-off-by: Ben Wilson <39283302+BenWilson2@users.noreply.github.com>
Co-authored-by: Serena Ruan <82044803+serena-ruan@users.noreply.github.com>
Signed-off-by: Joseph Mocker <mock@fakebelieve.org>