Make the Python SDK a namespace package (#688)
**PR Checklist**

- [X] A description of the changes is added to the description of this
PR.
- [ ] If there is a related issue, make sure it is linked to this PR.
- [ ] If you've fixed a bug or added code that should be tested, add
tests!
- [ ] If you've added or modified a feature, documentation in `docs` is
updated

**Description of changes**

This PR makes modifications to the generated Python SDK code so that it
conforms to the PEP requirements for inferred shared namespaces
https://packaging.python.org/en/latest/guides/packaging-namespace-packages/#legacy-namespace-packages
.
In order to accomplish this, the sbt build script has been updated to
create the proper directory structure, remove the root `__init__.py`
from the `unitycatalog.*` namespace, and create the distributable
`README.md` for PyPI, the `pyproject.toml` file for building the
deployable artifact, and the `setup.py` file for local developer-mode
installation of the package.

---------

Signed-off-by: Ben Wilson <benjamin.wilson@databricks.com>
Signed-off-by: Ben Wilson <39283302+BenWilson2@users.noreply.github.com>
Signed-off-by: Joseph Mocker <mock@fakebelieve.org>