Send non empty data as part of expired cookie in the header (#750)
**PR Checklist**

- [ ] A description of the changes is added to the description of this
PR.
- [ ] If there is a related issue, make sure it is linked to this PR.
- [x] If you've fixed a bug or added code that should be tested, add
tests!
- [ ] If you've added or modified a feature, documentation in `docs` is
updated

**Description of changes**

Logout endpoint was implemented to expire cookie as part of the task
https://github.com/unitycatalog/unitycatalog/pull/593 and changes are
pushed as part of PR
https://github.com/unitycatalog/unitycatalog/pull/678 but when the
endpoint was called it gets hung indefinitely , On further investigation
non empty data needs to be send as part of the header when we expire the
cookie. This PR has the fix for the same.

---------

Signed-off-by: sudharshanraja-db <sudharshan.rajan@databricks.com>
(cherry picked from commit 737a07a971da752912cfbea8e7cfac9b2c210207)