Move away from singleton pattern for properties and repositories  (#847)
**PR Checklist**

- [ ] A description of the changes is added to the description of this
PR.
- [ ] If there is a related issue, make sure it is linked to this PR.
- [ ] If you've fixed a bug or added code that should be tested, add
tests!
- [ ] If you've added or modified a feature, documentation in `docs` is
updated

**Description of changes**

<!-- Please state what you've changed and how it might affect the users.
-->
Currently, the properties and repositories classes are singleton and not
tied to a UnityCatalogServer object. This makes it difficult to
configure the server without using file based properties or system env
variables (also difficult to have multiple instances of the server for
tests). These changes uses dependency injection across classes to help
tie objects of these classes together instead of depending on the
singleton pattern.

---------

Signed-off-by: Vikrant Puppala <vikrant.puppala@databricks.com>