Add a function cache invalidator to the UnityCatalogFunctionsClient (#853)
**PR Checklist**

- [X] A description of the changes is added to the description of this
PR.
- [ ] If there is a related issue, make sure it is linked to this PR.
- [ ] If you've fixed a bug or added code that should be tested, add
tests!
- [ ] If you've added or modified a feature, documentation in `docs` is
updated

**Description of changes**

Adds a public method to the UnityCatalogFunctionsClient to allow for
invalidation of the execution cache for development purposes. If a
function has been overwritten or modified in the current REPL scope,
updates to a function will not reflect until either a new client is
initialized or the REPL session has been terminated. It's an annoying
devex that's fixed with the addition of this API.

---------

Signed-off-by: Ben Wilson <benjamin.wilson@databricks.com>