Fix serverless session resumption (#918)
**PR Checklist**

- [X] A description of the changes is added to the description of this
PR.
- [ ] If there is a related issue, make sure it is linked to this PR.
- [ ] If you've fixed a bug or added code that should be tested, add
tests!
- [ ] If you've added or modified a feature, documentation in `docs` is
updated

**Description of changes**

This PR forces the SparkSession (triggered by the call to
`set_default_spark_session` and the call to the builder to reconnect to
the serverless instance in the case of session or cluster invalidation)
to rebuild the spark session in accordance with recommendations of the
serverless team. Also, in the event that the retry logic is exhausted,
provide guidance for notebook users on how they can quickly get back up
and running by detaching and reattaching their notebook to an active
serverless cluster instance.

---------

Signed-off-by: Ben Wilson <benjamin.wilson@databricks.com>