Local exec (#925)
**PR Checklist**

- [X] A description of the changes is added to the description of this
PR.
- [ ] If there is a related issue, make sure it is linked to this PR.
- [ ] If you've fixed a bug or added code that should be tested, add
tests!
- [ ] If you've added or modified a feature, documentation in `docs` is
updated

**Description of changes**

Adds a local sandbox mode for execution that can protect the running
REPL context (and underlying VM) from computationally complex function
executions and limits the available libraries that can be utilized
within the function execution to provide a modicum of security as
compared to raw execution of a callable in the main process.

---------

Signed-off-by: Ben Wilson <benjamin.wilson@databricks.com>
Signed-off-by: Ben Wilson <39283302+BenWilson2@users.noreply.github.com>