Propagate Workspace Client Config to DB Connect (#930)
**PR Checklist**

- [x] A description of the changes is added to the description of this
PR.
- [ ] If there is a related issue, make sure it is linked to this PR.
- [ ] If you've fixed a bug or added code that should be tested, add
tests!
- [ ] If you've added or modified a feature, documentation in `docs` is
updated

**Description of changes**
In this PR, if a client is defined, the config is propagated to the
Databricks Session. This is done by setting the `sdkConfig` parameter.
However this parameter cannot be set when serverless is set to True. So
we first alter the config and set serverless to auto, which has the same
effect as setting `serverless` to True on Databricks Session.

<!-- Please state what you've changed and how it might affect the users.
-->

---------

Signed-off-by: aravind-segu <aravind.segu@databricks.com>