Set default client in get_uc_function_client (#940)
**PR Checklist**

- [ ] A description of the changes is added to the description of this
PR.
- [ ] If there is a related issue, make sure it is linked to this PR.
- [ ] If you've fixed a bug or added code that should be tested, add
tests!
- [ ] If you've added or modified a feature, documentation in `docs` is
updated

**Description of changes**
We don't need a warning for get_uc_function_client when it's none. It's
also confusing when it's used in validate_or_set_default_client.
This PR also moves the logic to set default client in
get_uc_function_client, so users don't need to create a
UCFunctionToolkit in order to trigger default client setup.

<!-- Please state what you've changed and how it might affect the users.
-->

---------

Signed-off-by: serena-ruan <serena.rxy@gmail.com>