Unity Catalog AI client handling of SQL NULL default parameters (#952)
**PR Checklist**

- [X] A description of the changes is added to the description of this
PR.
- [ ] If there is a related issue, make sure it is linked to this PR.
- [ ] If you've fixed a bug or added code that should be tested, add
tests!
- [ ] If you've added or modified a feature, documentation in `docs` is
updated

**Description of changes**

Currently, functions that have parameter defaults with SQL "NULL" as the
default value fail validation checks with a very difficult to
troubleshoot error message from pydantic validators.
This PR introduces the fix for SQL Null handling and establishes an
error handler to make it easier to understand why a Validator error
happens.
However, the larger-scale wrapping of validator checks elsewhere in the
code base would make a full replacement be many thousands of lines long,
so will leave usage of the decorator and the function handler to later
PRs.

---------

Signed-off-by: Ben Wilson <benjamin.wilson@databricks.com>