Fix time-travel on Delta tables using SQL (#995)
**Description of changes**

[UCSingleCatalog](https://github.com/unitycatalog/unitycatalog/blob/b141f19665c0c88ed6b72a04149881690ac29dcd/connectors/spark/src/main/scala/io/unitycatalog/spark/UCSingleCatalog.scala)
is missing overloads for `loadTable` with time-travel arguments
(version/timestamp). This effectively prevents time-travelling on Spark
+ Delta tables using SQL - Scala/python are working without this because
delta-spark resolves the table itself in that case without delegating to
Spark.

Note: this change doesn't add similar overloads to
[UCProxy](https://github.com/unitycatalog/unitycatalog/blob/b141f19665c0c88ed6b72a04149881690ac29dcd/connectors/spark/src/main/scala/io/unitycatalog/spark/UCSingleCatalog.scala#L205C15-L205C22),
so time-travel likely still won't work when a DeltaCatalog
implementation isn't available. I'm not actually sure how to properly
implement `loadTable` with time-travel there, so leaving it out of this
PR.

**Testing**
- Added tests for time-travel on a Delta table using SQL/Scala APIs in
`TableReadWriteTest`

Signed-off-by: Johan Lasperas <johan.lasperas@databricks.com>