[0.3] Add real production credential renwal integration test. (#1119)
**PR Checklist**

- [x] A description of the changes is added to the description of this
PR.
- [x] If there is a related issue, make sure it is linked to this PR.
- [x] If you've fixed a bug or added code that should be tested, add
tests!
- [ ] If you've added or modified a feature, documentation in `docs` is
updated

**Description of changes**

For this issue,
https://github.com/unitycatalog/unitycatalog/issues/1092. Since the
credential may be expired for 30minutes, so in this PR, I designed a
streaming job which will redirect the stream into the designated table
in 40 min. And if all the rows are correct, and the streaming job is not
been broken, then the credential renewal should be fine.

---------

Signed-off-by: Zheng Hu <openinx@gmail.com>