RetryHttpClient to enable Retry for UC (#1145)
**PR Checklist**

- [ ] A description of the changes is added to the description of this
PR.
- [ ] If there is a related issue, make sure it is linked to this PR.
- [ ] If you've fixed a bug or added code that should be tested, add
tests!
- [ ] If you've added or modified a feature, documentation in `docs` is
updated

**Description of changes**

Instead of wrapping each type of Api such as TempCredsApi or TablesApi
with a "Retryable" version, we instead create our own
RetryableHttpClient and retry at the http level. This way, new Apis just
need to pass the necessary configuration to the ApiClientFactory, which
creates a RetryingApiClient, which then creates a RetryingHttpClient.
The user may pass in a custom interceptor -- a function that defines
under which error codes, status codes, and exception they would want the
ApiClient to retry.

<!-- Please state what you've changed and how it might affect the users.
-->

---------

Signed-off-by: TimothyW553 <timothy.art@gmail.com>
Signed-off-by: Timothy Wang <timothy.art@gmail.com>