[Backport branch-0.3] allow object storage to be configured by env vars (#1150)
Backport of #1146 to `branch-0.3`.

---

**PR Checklist**

- [x] A description of the changes is added to the description of this
PR.
- [ ] If there is a related issue, make sure it is linked to this PR.
- [ ] If you've fixed a bug or added code that should be tested, add
tests!
- [ ] If you've added or modified a feature, documentation in `docs` is
updated

**Description of changes**

<!-- Please state what you've changed and how it might affect the users.
-->

It's currently possible to use environment variables to configure server
properties, however that's not the case for object storage
configurations, as it seems they are being fetched directly from the
properties files and not being fetched through the configuration
precedence, this PR unify that logic with the rest of the configurations
in server properteis

Signed-off-by: 1b5d <8110504+1b5d@users.noreply.github.com>
Signed-off-by: Tathagata Das <tathagata.das1565@gmail.com>
Co-authored-by: Abed <8110504+1b5d@users.noreply.github.com>