Allow staging table creation with the same name.
This is to fix a bug that UC didn't allow a 2nd staging table to be created using the same name. However this is actually a valid use case. When the table creation failed after the staging table is created by before the table creation, the dangling staging table would prevent retrying effort to create the table using the same name again. We have to allow the staging table to be created again using the same name to get it through. The new staging table will have new ID and location so it won't be a problem anyway.

Signed-off-by: Yi Li <yi.li@databricks.com>