Allow staging table creation with the same name. (#1230)
**PR Checklist**

- [X] A description of the changes is added to the description of this
PR.
- [X] If there is a related issue, make sure it is linked to this PR.
- [ ] If you've fixed a bug or added code that should be tested, add
tests!
- [ ] If you've added or modified a feature, documentation in `docs` is
updated

**Description of changes**

This is to fix a bug that UC didn't allow a 2nd staging table to be
created using the same name. However this is actually a valid use case.
When the table creation failed after the staging table is created but
before the table creation, the dangling staging table would prevent
retrying effort to create the table using the same name again. We have
to allow the staging table to be created again using the same name to
get it through. The new staging table will have new ID and location so
it won't be a problem anyway.

**Related issue**
https://github.com/unitycatalog/unitycatalog/issues/1143

Signed-off-by: Yi Li <yi.li@databricks.com>