[Backport branch-0.3] [java client] Handle unknown enum values gracefully in the java client (#1247)
Backport of #1177 to `branch-0.3`.

---

**PR Checklist**

- [x] A description of the changes is added to the description of this
PR.
- [ ] If there is a related issue, make sure it is linked to this PR.
- [x] If you've fixed a bug or added code that should be tested, add
tests!
- [ ] If you've added or modified a feature, documentation in `docs` is
updated

**Description of changes**

For various enums generated from the open api spec, its possible that an
implementation of the spec may have more values than the ones defined in
the spec. so its better to make the client not fail on encountaring such
unknown values in RPC responses. Rather it should handle gracefully to
produce the enum value of "UNKNOWN".

Signed-off-by: Tathagata Das <tathagata.das1565@gmail.com>