[0.3] Validation for ServerProperties. (#1271)
**PR Checklist**

- [x] A description of the changes is added to the description of this
PR.
- [ ] If there is a related issue, make sure it is linked to this PR.
- [ ] If you've fixed a bug or added code that should be tested, add
tests!
- [ ] If you've added or modified a feature, documentation in `docs` is
updated

**Description of changes**
Validation for ServerProperties.

ServerProperties is implemented as Map<String, String> to provide server
the essential settings. Before this PR if a invalid value is set for a
property it will only be checked and fail when that settings is actually
used. This PR would proactively validate property values to report
errors and crash when the server starts.

Signed-off-by: Yi Li <yi.li@databricks.com>