[0.3] Improve the abstracted unitycatalog-client API.
**PR Checklist**

- [x] A description of the changes is added to the description of this
PR.
- [x] If there is a related issue, make sure it is linked to this PR.
- [x] If you've fixed a bug or added code that should be tested, add
tests!
- [x] If you've added or modified a feature, documentation in `docs` is
updated

**Description of changes**

The related issue is :
https://github.com/unitycatalog/unitycatalog/issues/1259

This PR is the follow up of this one:
https://github.com/unitycatalog/unitycatalog/pull/1261

In this PR I did the following things:

* rename the `appVersion` to `addAppVersion`, which can append the app
version for multiple calls.
* Rename the `FixedTokenProvider` to `StaticTokenProvider`,which make
more sense.
* Add the `initializate(Map<String, String> configs>` in the abstracted
TokenProvider interface to initialize the provider instance. The core
problem is: if we use the abstraction way to initialize the provider,
then we have to introduce the `auth.type`, which can be `static`,
`oauth`, `xxxx.CustomizedTokenProvider`. With this, the abstracted
`TokenProvider#createFromConfigs` can dynamically load the provider
instance from the `Map<String, String>`. Otherwise, the `TokenProvider`
can still have no abstracted approach to load the instance, because it
doesnot know which to load. Another point is: we've already have
released the version, which support `uri` and `token` , so we will use
the `auth.type=static` by default if people did not specify the
`auth.type`. With this way, we can be compatible to the old version.
* Removed the original customized ApiClient template generation, since
currently we can customize the requestInteceptor directly to inject
those `User-Agent` information.

---------

Signed-off-by: Zheng Hu <openinx@gmail.com>
Co-authored-by: Tathagata Das <tathagata.das1565@gmail.com>
Signed-off-by: Zheng Hu <openinx@gmail.com>